Fixed training process. Now the PINN is trained on each batch of ODEs and its corresponding series coefficients.
Fixed the training process so that the PINN is trained seperately on
each batch of series coefficients. Right now there is an error:

ERROR: LoadError: Can't differentiate foreigncall expression
$(Expr(:foreigncall, :(:jl_toplevel_eval_
in), Any, svec(Any, Any), 0, :(:ccall), %2, %3)).
You might want to check the Zygote limitations documentation.
https://fluxml.ai/Zygote.jl/latest/limitations

I believe this has to do with how we are parsing the alpha matrix (the
ODE embedding) from a string to a matrix.