Trying to get the PINN to memorize some solutions
We added some weights to the pde_loss and the bcs_loss. When you set
both to zero and ask the pinn to recover a solution it has been trained
on, it can recover that function. However, when you set both of these
weights to be non-zero then the PINN performs worse. I also did an
experiment with a training dataset and a benchmark dataset. When both
weights are active the model performs badly. When one or the other is
non-zero it performs pretty well and can pick up on the growth of the
exponential.