Create graphs to plot the iteration between each seperate loss value
Created a function that stores the loss value for each optimization
step. Seperated the loss functions from the PINN file to a loss_funcs
file.

There is an issue though that the global loss function is where we call
the create csv file. It is incredible slow and I need to find a work
around.